fixed ran for retry
