+++ internal/testing/config/config.go
@@ -1,0 +1,6 @@
+package config
+
+const (
+	LockRedisDB = iota
+	RatelimitRedisDB
+)
