+++ lock/locker.go
@@ -2,7 +2,6 @@ package lock
 
 import (
 	"context"
-	"time"
 )
 
 type Locker interface {
@@ -11,21 +10,4 @@ type Locker interface {
 
 type Unlocker interface {
 	Unlock() error
-}
-
-type Options struct {
-	MaxRetry     int
-	RetryDurMin  time.Duration
-	RetryDurMax  time.Duration
-	RetryRandMax time.Duration
-}
-
-func (o *Options) SetDefaults() {
-	if o == nil {
-		return
-	}
-
-	if o.RetryRandMax == 0 {
-		o.RetryRandMax = DefaultRetryRandMax
-	}
 }
